Rewrote tests for map matching prod_data when using realistic data
A bug in the tests "test_prod_data_list_authenticated" and
"test_prod_data_list_authenticated_staff" produces unwantetd output in the console. The tests
themselve are running fine, but the test data in the PostProductionDataTest setUp self.data are
clutering the console. The self.data that passes the tests are printed out, which is undesirable.